[RESTEASY-2012] Getting rid of old Apache HttpClient 4.0 support, defaulting to closing httpclient when the engine is closed
