+++ resteasy-links/src/test/java/org/jboss/resteasy/links/test/TestLinkIds.java
@@ -4,11 +4,11 @@ import static org.jboss.resteasy.test.TestPortProvider.generateBaseUrl;
 
 import java.util.concurrent.TimeUnit;
 
-import org.apache.http.client.HttpClient;
+import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClientBuilder;
 import org.jboss.resteasy.client.jaxrs.ResteasyClientBuilder;
 import org.jboss.resteasy.client.jaxrs.ResteasyWebTarget;
-import org.jboss.resteasy.client.jaxrs.engines.ApacheHttpClient4Engine;
+import org.jboss.resteasy.client.jaxrs.engines.ApacheHttpClientEngine;
 import org.jboss.resteasy.core.Dispatcher;
 import org.jboss.resteasy.links.RESTServiceDiscovery;
 import org.jboss.resteasy.links.RESTServiceDiscovery.AtomLink;
@@ -38,7 +38,7 @@ public class TestLinkIds
       POJOResourceFactory noDefaults = new POJOResourceFactory(IDServiceTestBean.class);
       dispatcher.getRegistry().addResourceFactory(noDefaults);
       httpClient = HttpClientBuilder.create().build();
-      ApacheHttpClient4Engine engine = new ApacheHttpClient4Engine(httpClient);
+      ApacheHttpClientEngine engine = ApacheHttpClientEngine.create(httpClient);
       url = generateBaseUrl();
       ResteasyWebTarget target = new ResteasyClientBuilder().httpEngine(engine).build().target(url);
       client = target.proxy(IDServiceTest.class);
@@ -54,7 +54,7 @@ public class TestLinkIds
 
 	private static String url;
 	private static IDServiceTest client;
-	private static HttpClient httpClient;
+	private static CloseableHttpClient httpClient;
 	
 	@SuppressWarnings("deprecation")
     @After
