+++ testsuite/integration-tests/src/test/java/org/jboss/resteasy/test/client/ClientExecutorShutdownTest.java
@@ -10,7 +10,8 @@ import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.logging.Logger;
 import org.jboss.resteasy.client.jaxrs.ResteasyClient;
 import org.jboss.resteasy.client.jaxrs.ResteasyClientBuilder;
-import org.jboss.resteasy.client.jaxrs.engines.ApacheHttpClient4Engine;
+import org.jboss.resteasy.client.jaxrs.engines.ApacheHttpClient43Engine;
+import org.jboss.resteasy.client.jaxrs.engines.ApacheHttpClientEngine;
 import org.jboss.resteasy.test.client.resource.ClientExecutorShutdownTestResource;
 import org.jboss.resteasy.util.HttpResponseCodes;
 import org.jboss.resteasy.utils.TestUtil;
@@ -60,7 +61,7 @@ public class ClientExecutorShutdownTest extends ClientTestBase{
      */
     @Test
     public void testApacheHttpClient4ExecutorNonSharedHttpClientFinalize() throws Throwable {
-        ApacheHttpClient4Engine engine = new ApacheHttpClient4Engine();
+        ApacheHttpClient43Engine engine = new ApacheHttpClient43Engine();
         ResteasyClient client = new ResteasyClientBuilder().httpEngine(engine).build();
         Response response = client.target(generateURL("/test")).request().post(null);
         Assert.assertEquals(HttpResponseCodes.SC_NO_CONTENT, response.getStatus());
@@ -84,7 +85,7 @@ public class ClientExecutorShutdownTest extends ClientTestBase{
      */
     @Test
     public void testApacheHttpClient4ExecutorNonSharedHttpClientClose() throws Throwable {
-        ApacheHttpClient4Engine engine = new ApacheHttpClient4Engine();
+        ApacheHttpClient43Engine engine = new ApacheHttpClient43Engine();
         ResteasyClient client = new ResteasyClientBuilder().httpEngine(engine).build();
         Response response = client.target(generateURL("/test")).request().post(null);
         Assert.assertEquals(HttpResponseCodes.SC_NO_CONTENT, response.getStatus());
@@ -109,7 +110,7 @@ public class ClientExecutorShutdownTest extends ClientTestBase{
     @Test
     public void testApacheHttpClient4ExecutorSharedHttpClientFinalize() throws Throwable {
         HttpClient httpClient = HttpClientBuilder.create().build();
-        ApacheHttpClient4Engine engine = new ApacheHttpClient4Engine(httpClient);
+        ApacheHttpClient43Engine engine = new ApacheHttpClient43Engine(httpClient, false);
         ResteasyClient client = new ResteasyClientBuilder().httpEngine(engine).build();
         Response response = client.target(generateURL("/test")).request().post(null);
         Assert.assertEquals(HttpResponseCodes.SC_NO_CONTENT, response.getStatus());
@@ -132,7 +133,7 @@ public class ClientExecutorShutdownTest extends ClientTestBase{
     @Test
     public void testApacheHttpClient4ExecutorSharedHttpClientClose() throws Throwable {
         HttpClient httpClient = HttpClientBuilder.create().build();
-        ApacheHttpClient4Engine engine = new ApacheHttpClient4Engine(httpClient);
+        ApacheHttpClient43Engine engine = new ApacheHttpClient43Engine(httpClient, false);
         ResteasyClient client = new ResteasyClientBuilder().httpEngine(engine).build();
         Response response = client.target(generateURL("/test")).request().post(null);
         Assert.assertEquals("Original httpclient and engine httpclient are not the same instance",
