+++ testsuite/integration-tests/src/test/java/org/jboss/resteasy/test/client/other/resource/CustomHttpClientEngineBuilder.java
@@ -8,7 +8,6 @@ import org.apache.http.impl.client.HttpClientBuilder;
 import org.jboss.resteasy.client.jaxrs.ClientHttpEngine;
 import org.jboss.resteasy.client.jaxrs.ClientHttpEngineBuilder43;
 import org.jboss.resteasy.client.jaxrs.engines.ApacheHttpClient43Engine;
-import org.jboss.resteasy.client.jaxrs.engines.factory.ApacheHttpClient4EngineFactory;
 
 import javax.net.ssl.HostnameVerifier;
 import javax.net.ssl.SSLContext;
@@ -24,8 +23,7 @@ public class CustomHttpClientEngineBuilder extends ClientHttpEngineBuilder43 {
                 .setDefaultRequestConfig(rcBuilder.build())
                 .setProxy(defaultProxy)
                 .disableContentCompression().build();
-        ApacheHttpClient43Engine engine = (ApacheHttpClient43Engine) ApacheHttpClient4EngineFactory.create(httpClient,
-                true);
+        ApacheHttpClient43Engine engine = new ApacheHttpClient43Engine(httpClient, true);
         engine.setResponseBufferSize(responseBufferSize);
         engine.setHostnameVerifier(verifier);
         // this may be null.  We can't really support this with Apache Client.
